requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pip==23.0.1
pytz==2023.3
djangorestframework_simplejwt==5.2.2
passlib==1.7.4
Pillow==9.5.0
django_utils_six==2.0